pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "teeshield"
version = "0.2.1"
description = "Security linter for MCP servers and AI agents -- 46 checks for vulnerabilities, malicious patterns, and description quality"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "TeeShield" },
]
keywords = ["mcp", "ai-agent", "security", "tool-firewall", "skill-rating", "static-analysis", "malware-detection", "sarif"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
ai = [
    "anthropic>=0.40",
    "openai>=1.0",
    "google-generativeai>=0.5",
]

[project.scripts]
teeshield = "teeshield.cli:main"
teeshield-server = "teeshield.server:run"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
